js/xpconnect/tests/mochitest/test_bug623437.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=623437
-->
<head>
  <title>Test for Bug 623437</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=623437">Mozilla Bug 623437</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 623437 **/
    var c = document.createElement("canvas").getContext("2d");
    var seenArcToFirst;
    var seenArc = false;
    var seenArcTo = false;
    for (var i in c) {
      if (i == "arc") {
        seenArc = true;
        if (!seenArcTo)
          seenArcToFirst = false;
      }
      if (i == "arcTo") {
        seenArcTo = true;
        if (!seenArc)
          seenArcToFirst = true;
      }
    }
    is(seenArc, true, "Should see arc");
    is(seenArcTo, true, "Should see arcTo");
    is(seenArcToFirst, true, "Should see arcTo before arc");
</script>
</pre>
</body>
</html>
